feat(location): location is ready for test   - added getCitiesList and getCountriesList methods;
