+++ routes/index.ts
@@ -11,7 +11,7 @@ export class Routes {
     require('./appriciations')(app);
     require('./upload-files')(app);
     require('./follow')(app);
-    require('./location')(app);
+    require('./location.route')(app);
     require('./genres')(app);
     require('./nodemailer.route')(app);
   }
